requirements.txt
pytest==6.1.1
flask-restplus==0.9.2
Flask-SQLAlchemy==2.1
werkzeug==0.16.1
